smite: move commitment tests to submodule
